+++ packages/carbon-graphs/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Changed
-  * Updated reflow to allow users to send multiple datasets per graph Instance.
+  * Updated reflow to allow users to send multiple datasets per graph instance.
 
 ## 2.16.0 - (December 16, 2020)